UPD: Refactored google oauth to make the function testable
